Split Config_spec.js into one file per method
- Config_getResource_spec.js
- Config_getClient_spec.js
- Config_fromFile_spec.js
